+++ models/demos/deepseek_v3_d_p/tests/pcc/test_reduce.py
@@ -159,11 +159,11 @@ def test_ttnn_reduce(
         dims=(0, None),
     )
     tt_indices = ttnn.from_torch(
-        torch_indices,
+        torch_indices.to(torch.int16),
         mesh_mapper=indices_mapper,
         layout=ttnn.ROW_MAJOR_LAYOUT,
         device=mesh_device,
-        dtype=ttnn.int32,
+        dtype=ttnn.uint16,
     )
     tt_expert_dispatch_table = ttnn.from_torch(
         expert_dispatch_table,
